Use addSpacing instead of insertSpacing
Prefer to use addSpacing.  If the index value is incorrect for
insertSpacing, Flame crashes with no helpful message.  Its very tricky
to debug.  addSpacing renders the same as insertSpacing.  Also, placing
an addSpacing where the space is to be added is simpler to read than
throwing insertSpacing wherever in the QBoxLayout and relying on
indexes.